+++ conda_recipe_manager/commands/rattler_bulk_build.py
@@ -28,6 +28,53 @@ RATTLER_ERROR_REGEX = re.compile(r"Error:\s+.*")
 # Timeout to halt operation
 DEFAULT_RATTLER_BUILD_TIMEOUT: Final[int] = 120
 
+# Common variables seen in Conda Build Config (CBC) files. As of `rattler-build v0.58.0+`, our integration tests will
+# erroneously fail if these are not provided. We could use a conda-forge or Anaconda-provided CBC file, but this method
+# ensures we are consistent across ecosystems represented in the tests AND prevents discrepancies with changing external
+# files.
+_RATTLER_BUILD_VARIANT_VARS: Final = [
+    "python=3.11",
+    "python=3.12",
+    "python=3.13",
+    "python_implementation=cpython",
+    "python_impl=cpython",
+    "numpy=2.4.2",
+    "blas_impl=openblas",
+    "c_compiler=gcc",
+    "c_stdlib=sysroot",
+    "cxx_compiler=gxx",
+    "cuda_compiler=cuda-nvcc",
+    "fortran_compiler=gfortran",
+    "m2w64_c_compiler=m2w64-toolchain",
+    "m2w64_cxx_compiler=m2w64-toolchain",
+    "m2w64_fortran_compiler=m2w64-toolchain",
+    "ucrt64_c_compiler=ucrt64-gcc-toolchain",
+    "ucrt64_cxx_compiler=ucrt64-gcc-toolchain",
+    "ucrt64_fortran_compiler=ucrt64-gcc-toolchain",
+    "rust_compiler=rust",
+    "rust_nightly_compiler=rust-nightly",
+    "rust_compiler_version=1.93.1",
+    "rust_nightly_compiler_version=1.92.0_2025-10-13",
+    "rust_gnu_compiler=rust-gnu",
+    "rust_gnu_compiler_version=1.93.1",
+    "VERBOSE_AT=V=1",
+    "VERBOSE_CM=VERBOSE=1",
+    "cran_mirror=https://cran.r-project.org",
+    "c_compiler_version=14.3.0",
+    "c_stdlib_version=2.28",
+    "cxx_compiler_version=14.3.0",
+    "cuda_compiler_version=12.4",
+    "fortran_compiler_version=14.3.0",
+    "clang_variant=clang",
+    "go_compiler=go-nocgo",
+    "go_compiler_version=1.21",
+    "cgo_compiler=go-cgo",
+    "cgo_compiler_version=1.21",
+    "r_base=4.3.1",
+    "r_version=4.3.1",
+    "r_implementation=r-base",
+]
+
 
 @dataclass
 class BuildResult:
@@ -39,24 +86,29 @@ class BuildResult:
     errors: list[str]
 
 
-def build_recipe(file: Path, path: Path, args: list[str]) -> tuple[str, BuildResult]:
+def _build_recipe(file: Path, path: Path, inject_vars: bool, args: list[str]) -> tuple[str, BuildResult]:
     """
     Helper function that performs the build operation for parallelizable execution. Logs rattler-build failures to
     STDERR.
 
     :param file: Recipe file to build
     :param path: Path argument provided by the user
+    :param inject_vars: Flag indicating if fake recipe variables should be injected into the build command. Used for
+        CRM's integration tests.
     :param args: List of arguments to provide whole-sale to rattler-build
     :returns: Tuple containing the key/value pairing that tracks the result of the build operation
     """
     cmd: list[str] = ["rattler-build", "build", "-r", str(file)]
+    if inject_vars:
+        for var in _RATTLER_BUILD_VARIANT_VARS:
+            cmd += ["--variant", var]
     cmd.extend(args)
     try:
         output: Final[subprocess.CompletedProcess[str]] = subprocess.run(
-            " ".join(cmd),
+            cmd,
             encoding="utf-8",
             capture_output=True,
-            shell=True,
+            shell=False,
             check=False,
             timeout=DEFAULT_RATTLER_BUILD_TIMEOUT,
         )
@@ -72,7 +124,7 @@ def build_recipe(file: Path, path: Path, args: list[str]) -> tuple[str, BuildRes
     )
 
 
-def create_debug_file(debug_log: Path, results: dict[str, BuildResult], error_histogram: dict[str, int]) -> None:
+def _create_debug_file(debug_log: Path, results: dict[str, BuildResult], error_histogram: dict[str, int]) -> None:
     """
     Generates a debug file containing an organized dump of all the recipes that got a particular error message.
 
@@ -136,9 +188,21 @@ def create_debug_file(debug_log: Path, results: dict[str, BuildResult], error_hi
     type=click.Path(exists=False, file_okay=True, dir_okay=False, path_type=Path),
     help="Dumps a large debug log to the file specified.",
 )
+# Added to support `rattler-build v0.58.0+`, which fails builds if JINJA variables are not defined. These variables
+# are commonly found in Conda Build Config files.
+@click.option(
+    "--inject-vars",
+    is_flag=True,
+    help="Injects variables for `rattler-build` so that integration tests don't fail on undefined variables.",
+)
 @click.pass_context
 def rattler_bulk_build(
-    ctx: click.Context, path: Path, min_success_rate: float, truncate: bool, debug_log: Optional[Path]
+    ctx: click.Context,
+    path: Path,
+    min_success_rate: float,
+    truncate: bool,
+    debug_log: Optional[Path],
+    inject_vars: bool,
 ) -> None:
     """
     Given a directory of feedstock repositories, performs multiple recipe builds using rattler-build.
@@ -160,7 +224,9 @@ def rattler_bulk_build(
     # Process recipes in parallel
     thread_pool_size: Final[int] = mp.cpu_count()
     with mp.Pool(thread_pool_size) as pool:
-        results = dict(pool.starmap(build_recipe, [(file, path, ctx.args) for file in files]))  # type: ignore[misc]
+        results = dict(
+            pool.starmap(_build_recipe, [(file, path, inject_vars, ctx.args) for file in files])  # type: ignore[misc]
+        )
 
     # Gather statistics
     total_recipes: Final[int] = len(files)
@@ -207,7 +273,7 @@ def rattler_bulk_build(
         final_output["recipes_with_build_error_code"] = recipes_with_errors
 
     if debug_log is not None:
-        create_debug_file(debug_log, results, error_histogram)
+        _create_debug_file(debug_log, results, error_histogram)
 
     print(json.dumps(final_output, indent=2))
     sys.exit(ExitCode.SUCCESS if percent_success >= min_success_rate else ExitCode.MISSED_SUCCESS_THRESHOLD)